# Remove roSingleSegment arena